requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.1